Terminada pantalla de asteroides y corregido cambios entre escenas y correcciones varias

+++ Dolphin Space Hunter/Assets/_Scripts/Shield_Recharge/ShieldScript.cs
@@ -15,7 +15,7 @@ public class ShieldScript : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        Shield = PlayerPrefs.GetFloat("Shield", 0f);
+        Shield = PlayerPrefs.GetInt("Shield", 0);
         currentAmmunition = PlayerPrefs.GetInt("Ammo", 0);
         maxAmmunition = PlayerPrefs.GetInt("MaxAmmo", 50);
         currentScore = PlayerPrefs.GetInt("PlayerScore", 0);
